WIP: Use Browserify to bundle Sinon.JS
