+++ lib/sinon/behavior.js
@@ -1,8 +1,5 @@
-/**
- * @depend ../sinon.js
- */
 /*jslint eqeqeq: false, onevar: false*/
-/*global module, require, sinon, process, setImmediate, setTimeout*/
+/*global module, require, process, setImmediate, setTimeout*/
 /**
  * Stub behavior
  *
@@ -14,319 +11,303 @@
  */
 "use strict";
 
-(function (sinon) {
-    var commonJSModule = typeof module !== 'undefined' && module.exports;
+var sinon = require("./sinon");
+var slice = Array.prototype.slice;
+var join = Array.prototype.join;
 
-    if (!sinon && commonJSModule) {
-        sinon = require("../sinon");
+var nextTick = (function () {
+    if (typeof process === "object" && typeof process.nextTick === "function") {
+        return process.nextTick;
+    } else if (typeof setImmediate === "function") {
+        return setImmediate;
+    } else {
+        return function (callback) {
+            setTimeout(callback, 0);
+        };
+    }
+})();
+
+function throwsException(error, message) {
+    if (typeof error == "string") {
+        this.exception = new Error(message || "");
+        this.exception.name = error;
+    } else if (!error) {
+        this.exception = new Error("Error");
+    } else {
+        this.exception = error;
     }
 
-    if (!sinon) {
-        return;
+    return this;
+}
+
+function getCallback(behavior, args) {
+    var callArgAt = behavior.callArgAt;
+
+    if (callArgAt < 0) {
+        var callArgProp = behavior.callArgProp;
+
+        for (var i = 0, l = args.length; i < l; ++i) {
+            if (!callArgProp && typeof args[i] == "function") {
+                return args[i];
+            }
+
+            if (callArgProp && args[i] &&
+                typeof args[i][callArgProp] == "function") {
+                return args[i][callArgProp];
+            }
+        }
+
+        return null;
     }
 
-    var slice = Array.prototype.slice;
-    var join = Array.prototype.join;
-    var proto;
+    return args[callArgAt];
+}
 
-    var nextTick = (function () {
-        if (typeof process === "object" && typeof process.nextTick === "function") {
-            return process.nextTick;
-        } else if (typeof setImmediate === "function") {
-            return setImmediate;
-        } else {
-            return function (callback) {
-                setTimeout(callback, 0);
-            };
-        }
-    })();
+function getCallbackError(behavior, func, args) {
+    if (behavior.callArgAt < 0) {
+        var msg;
 
-    function throwsException(error, message) {
-        if (typeof error == "string") {
-            this.exception = new Error(message || "");
-            this.exception.name = error;
-        } else if (!error) {
-            this.exception = new Error("Error");
+        if (behavior.callArgProp) {
+            msg = sinon.functionName(behavior.stub) +
+                " expected to yield to '" + behavior.callArgProp +
+                "', but no object with such a property was passed.";
         } else {
-            this.exception = error;
+            msg = sinon.functionName(behavior.stub) +
+                " expected to yield, but no callback was passed.";
         }
+
+        if (args.length > 0) {
+            msg += " Received [" + join.call(args, ", ") + "]";
+        }
+
+        return msg;
+    }
+
+    return "argument at index " + behavior.callArgAt + " is not a function: " + func;
+}
+
+function callCallback(behavior, args) {
+    if (typeof behavior.callArgAt == "number") {
+        var func = getCallback(behavior, args);
+
+        if (typeof func != "function") {
+            throw new TypeError(getCallbackError(behavior, func, args));
+        }
+
+        if (behavior.callbackAsync) {
+            nextTick(function() {
+                func.apply(behavior.callbackContext, behavior.callbackArguments);
+            });
+        } else {
+            func.apply(behavior.callbackContext, behavior.callbackArguments);
+        }
+    }
+}
+
+var proto = {
+    create: function(stub) {
+        var behavior = sinon.extend({}, proto);
+        delete behavior.create;
+        behavior.stub = stub;
+
+        return behavior;
+    },
+
+    isPresent: function() {
+        return (typeof this.callArgAt == 'number' ||
+                this.exception ||
+                typeof this.returnArgAt == 'number' ||
+                this.returnThis ||
+                this.returnValueDefined);
+    },
+
+    invoke: function(context, args) {
+        callCallback(this, args);
+
+        if (this.exception) {
+            throw this.exception;
+        } else if (typeof this.returnArgAt == 'number') {
+            return args[this.returnArgAt];
+        } else if (this.returnThis) {
+            return context;
+        }
+
+        return this.returnValue;
+    },
+
+    onCall: function(index) {
+        return this.stub.onCall(index);
+    },
+
+    onFirstCall: function() {
+        return this.stub.onFirstCall();
+    },
+
+    onSecondCall: function() {
+        return this.stub.onSecondCall();
+    },
+
+    onThirdCall: function() {
+        return this.stub.onThirdCall();
+    },
+
+    withArgs: function(/* arguments */) {
+        throw new Error('Defining a stub by invoking "stub.onCall(...).withArgs(...)" is not supported. ' +
+                        'Use "stub.withArgs(...).onCall(...)" to define sequential behavior for calls with certain arguments.');
+    },
+
+    callsArg: function callsArg(pos) {
+        if (typeof pos != "number") {
+            throw new TypeError("argument index is not number");
+        }
+
+        this.callArgAt = pos;
+        this.callbackArguments = [];
+        this.callbackContext = undefined;
+        this.callArgProp = undefined;
+        this.callbackAsync = false;
+
+        return this;
+    },
+
+    callsArgOn: function callsArgOn(pos, context) {
+        if (typeof pos != "number") {
+            throw new TypeError("argument index is not number");
+        }
+        if (typeof context != "object") {
+            throw new TypeError("argument context is not an object");
+        }
+
+        this.callArgAt = pos;
+        this.callbackArguments = [];
+        this.callbackContext = context;
+        this.callArgProp = undefined;
+        this.callbackAsync = false;
+
+        return this;
+    },
+
+    callsArgWith: function callsArgWith(pos) {
+        if (typeof pos != "number") {
+            throw new TypeError("argument index is not number");
+        }
+
+        this.callArgAt = pos;
+        this.callbackArguments = slice.call(arguments, 1);
+        this.callbackContext = undefined;
+        this.callArgProp = undefined;
+        this.callbackAsync = false;
+
+        return this;
+    },
+
+    callsArgOnWith: function callsArgWith(pos, context) {
+        if (typeof pos != "number") {
+            throw new TypeError("argument index is not number");
+        }
+        if (typeof context != "object") {
+            throw new TypeError("argument context is not an object");
+        }
+
+        this.callArgAt = pos;
+        this.callbackArguments = slice.call(arguments, 2);
+        this.callbackContext = context;
+        this.callArgProp = undefined;
+        this.callbackAsync = false;
+
+        return this;
+    },
+
+    yields: function () {
+        this.callArgAt = -1;
+        this.callbackArguments = slice.call(arguments, 0);
+        this.callbackContext = undefined;
+        this.callArgProp = undefined;
+        this.callbackAsync = false;
+
+        return this;
+    },
+
+    yieldsOn: function (context) {
+        if (typeof context != "object") {
+            throw new TypeError("argument context is not an object");
+        }
+
+        this.callArgAt = -1;
+        this.callbackArguments = slice.call(arguments, 1);
+        this.callbackContext = context;
+        this.callArgProp = undefined;
+        this.callbackAsync = false;
+
+        return this;
+    },
+
+    yieldsTo: function (prop) {
+        this.callArgAt = -1;
+        this.callbackArguments = slice.call(arguments, 1);
+        this.callbackContext = undefined;
+        this.callArgProp = prop;
+        this.callbackAsync = false;
+
+        return this;
+    },
+
+    yieldsToOn: function (prop, context) {
+        if (typeof context != "object") {
+            throw new TypeError("argument context is not an object");
+        }
+
+        this.callArgAt = -1;
+        this.callbackArguments = slice.call(arguments, 2);
+        this.callbackContext = context;
+        this.callArgProp = prop;
+        this.callbackAsync = false;
+
+        return this;
+    },
+
+
+    "throws": throwsException,
+    throwsException: throwsException,
+
+    returns: function returns(value) {
+        this.returnValue = value;
+        this.returnValueDefined = true;
+
+        return this;
+    },
+
+    returnsArg: function returnsArg(pos) {
+        if (typeof pos != "number") {
+            throw new TypeError("argument index is not number");
+        }
+
+        this.returnArgAt = pos;
+
+        return this;
+    },
+
+    returnsThis: function returnsThis() {
+        this.returnThis = true;
 
         return this;
     }
+};
 
-    function getCallback(behavior, args) {
-        var callArgAt = behavior.callArgAt;
-
-        if (callArgAt < 0) {
-            var callArgProp = behavior.callArgProp;
-
-            for (var i = 0, l = args.length; i < l; ++i) {
-                if (!callArgProp && typeof args[i] == "function") {
-                    return args[i];
-                }
-
-                if (callArgProp && args[i] &&
-                    typeof args[i][callArgProp] == "function") {
-                    return args[i][callArgProp];
-                }
-            }
-
-            return null;
-        }
-
-        return args[callArgAt];
+// create asynchronous versions of callsArg* and yields* methods
+for (var method in proto) {
+    // need to avoid creating anotherasync versions of the newly added async methods
+    if (proto.hasOwnProperty(method) &&
+        method.match(/^(callsArg|yields)/) &&
+        !method.match(/Async/)) {
+        proto[method + 'Async'] = (function (syncFnName) {
+            return function () {
+                var result = this[syncFnName].apply(this, arguments);
+                this.callbackAsync = true;
+                return result;
+            };
+        })(method);
     }
+}
 
-    function getCallbackError(behavior, func, args) {
-        if (behavior.callArgAt < 0) {
-            var msg;
-
-            if (behavior.callArgProp) {
-                msg = sinon.functionName(behavior.stub) +
-                    " expected to yield to '" + behavior.callArgProp +
-                    "', but no object with such a property was passed.";
-            } else {
-                msg = sinon.functionName(behavior.stub) +
-                    " expected to yield, but no callback was passed.";
-            }
-
-            if (args.length > 0) {
-                msg += " Received [" + join.call(args, ", ") + "]";
-            }
-
-            return msg;
-        }
-
-        return "argument at index " + behavior.callArgAt + " is not a function: " + func;
-    }
-
-    function callCallback(behavior, args) {
-        if (typeof behavior.callArgAt == "number") {
-            var func = getCallback(behavior, args);
-
-            if (typeof func != "function") {
-                throw new TypeError(getCallbackError(behavior, func, args));
-            }
-
-            if (behavior.callbackAsync) {
-                nextTick(function() {
-                    func.apply(behavior.callbackContext, behavior.callbackArguments);
-                });
-            } else {
-                func.apply(behavior.callbackContext, behavior.callbackArguments);
-            }
-        }
-    }
-
-    proto = {
-        create: function(stub) {
-            var behavior = sinon.extend({}, sinon.behavior);
-            delete behavior.create;
-            behavior.stub = stub;
-
-            return behavior;
-        },
-
-        isPresent: function() {
-            return (typeof this.callArgAt == 'number' ||
-                    this.exception ||
-                    typeof this.returnArgAt == 'number' ||
-                    this.returnThis ||
-                    this.returnValueDefined);
-        },
-
-        invoke: function(context, args) {
-            callCallback(this, args);
-
-            if (this.exception) {
-                throw this.exception;
-            } else if (typeof this.returnArgAt == 'number') {
-                return args[this.returnArgAt];
-            } else if (this.returnThis) {
-                return context;
-            }
-
-            return this.returnValue;
-        },
-
-        onCall: function(index) {
-            return this.stub.onCall(index);
-        },
-
-        onFirstCall: function() {
-            return this.stub.onFirstCall();
-        },
-
-        onSecondCall: function() {
-            return this.stub.onSecondCall();
-        },
-
-        onThirdCall: function() {
-            return this.stub.onThirdCall();
-        },
-
-        withArgs: function(/* arguments */) {
-            throw new Error('Defining a stub by invoking "stub.onCall(...).withArgs(...)" is not supported. ' +
-                            'Use "stub.withArgs(...).onCall(...)" to define sequential behavior for calls with certain arguments.');
-        },
-
-        callsArg: function callsArg(pos) {
-            if (typeof pos != "number") {
-                throw new TypeError("argument index is not number");
-            }
-
-            this.callArgAt = pos;
-            this.callbackArguments = [];
-            this.callbackContext = undefined;
-            this.callArgProp = undefined;
-            this.callbackAsync = false;
-
-            return this;
-        },
-
-        callsArgOn: function callsArgOn(pos, context) {
-            if (typeof pos != "number") {
-                throw new TypeError("argument index is not number");
-            }
-            if (typeof context != "object") {
-                throw new TypeError("argument context is not an object");
-            }
-
-            this.callArgAt = pos;
-            this.callbackArguments = [];
-            this.callbackContext = context;
-            this.callArgProp = undefined;
-            this.callbackAsync = false;
-
-            return this;
-        },
-
-        callsArgWith: function callsArgWith(pos) {
-            if (typeof pos != "number") {
-                throw new TypeError("argument index is not number");
-            }
-
-            this.callArgAt = pos;
-            this.callbackArguments = slice.call(arguments, 1);
-            this.callbackContext = undefined;
-            this.callArgProp = undefined;
-            this.callbackAsync = false;
-
-            return this;
-        },
-
-        callsArgOnWith: function callsArgWith(pos, context) {
-            if (typeof pos != "number") {
-                throw new TypeError("argument index is not number");
-            }
-            if (typeof context != "object") {
-                throw new TypeError("argument context is not an object");
-            }
-
-            this.callArgAt = pos;
-            this.callbackArguments = slice.call(arguments, 2);
-            this.callbackContext = context;
-            this.callArgProp = undefined;
-            this.callbackAsync = false;
-
-            return this;
-        },
-
-        yields: function () {
-            this.callArgAt = -1;
-            this.callbackArguments = slice.call(arguments, 0);
-            this.callbackContext = undefined;
-            this.callArgProp = undefined;
-            this.callbackAsync = false;
-
-            return this;
-        },
-
-        yieldsOn: function (context) {
-            if (typeof context != "object") {
-                throw new TypeError("argument context is not an object");
-            }
-
-            this.callArgAt = -1;
-            this.callbackArguments = slice.call(arguments, 1);
-            this.callbackContext = context;
-            this.callArgProp = undefined;
-            this.callbackAsync = false;
-
-            return this;
-        },
-
-        yieldsTo: function (prop) {
-            this.callArgAt = -1;
-            this.callbackArguments = slice.call(arguments, 1);
-            this.callbackContext = undefined;
-            this.callArgProp = prop;
-            this.callbackAsync = false;
-
-            return this;
-        },
-
-        yieldsToOn: function (prop, context) {
-            if (typeof context != "object") {
-                throw new TypeError("argument context is not an object");
-            }
-
-            this.callArgAt = -1;
-            this.callbackArguments = slice.call(arguments, 2);
-            this.callbackContext = context;
-            this.callArgProp = prop;
-            this.callbackAsync = false;
-
-            return this;
-        },
-
-
-        "throws": throwsException,
-        throwsException: throwsException,
-
-        returns: function returns(value) {
-            this.returnValue = value;
-            this.returnValueDefined = true;
-
-            return this;
-        },
-
-        returnsArg: function returnsArg(pos) {
-            if (typeof pos != "number") {
-                throw new TypeError("argument index is not number");
-            }
-
-            this.returnArgAt = pos;
-
-            return this;
-        },
-
-        returnsThis: function returnsThis() {
-            this.returnThis = true;
-
-            return this;
-        }
-    };
-
-    // create asynchronous versions of callsArg* and yields* methods
-    for (var method in proto) {
-        // need to avoid creating anotherasync versions of the newly added async methods
-        if (proto.hasOwnProperty(method) &&
-            method.match(/^(callsArg|yields)/) &&
-            !method.match(/Async/)) {
-            proto[method + 'Async'] = (function (syncFnName) {
-                return function () {
-                    var result = this[syncFnName].apply(this, arguments);
-                    this.callbackAsync = true;
-                    return result;
-                };
-            })(method);
-        }
-    }
-
-    if (commonJSModule) {
-        module.exports = proto;
-    } else {
-        sinon.behavior = proto;
-    }
-}(typeof sinon == "object" && sinon || null));
+module.exports = proto;
