+++ lib/sinon/collection.js
@@ -1,10 +1,5 @@
-/**
- * @depend ../sinon.js
- * @depend stub.js
- * @depend mock.js
- */
 /*jslint eqeqeq: false, onevar: false, forin: true*/
-/*global module, require, sinon*/
+/*global module, require*/
 /**
  * Collections of stubs, spies and mocks.
  *
@@ -15,53 +10,31 @@
  */
 "use strict";
 
-(function (sinon) {
-    var commonJSModule = typeof module !== 'undefined' && module.exports;
-    var push = [].push;
-    var hasOwnProperty = Object.prototype.hasOwnProperty;
+var sinon = require("./sinon");
+var hasOwnProperty = Object.prototype.hasOwnProperty;
 
-    if (!sinon && commonJSModule) {
-        sinon = require("../sinon");
-    }
-
-    if (!sinon) {
-        return;
-    }
-
-    function getFakes(fakeCollection) {
-        if (!fakeCollection.fakes) {
-            fakeCollection.fakes = [];
-        }
-
-        return fakeCollection.fakes;
-    }
-
-    function each(fakeCollection, method) {
-        var fakes = getFakes(fakeCollection);
-
-        for (var i = 0, l = fakes.length; i < l; i += 1) {
-            if (typeof fakes[i][method] == "function") {
-                fakes[i][method]();
-            }
+function each(fakes, method) {
+    for (var i = 0, l = fakes.length; i < l; i += 1) {
+        if (typeof fakes[i][method] == "function") {
+            fakes[i][method]();
         }
     }
+}
 
-    function compact(fakeCollection) {
-        var fakes = getFakes(fakeCollection);
-        var i = 0;
-        while (i < fakes.length) {
-          fakes.splice(i, 1);
-        }
-    }
+module.exports = function (globalSinon) {
+    return {
 
-    var collection = {
         verify: function resolve() {
-            each(this, "verify");
+            if (this.fakes) {
+                each(this.fakes, "verify");
+            }
         },
 
         restore: function restore() {
-            each(this, "restore");
-            compact(this);
+            if (this.fakes) {
+                each(this.fakes, "restore");
+                this.fakes.length = 0;
+            }
         },
 
         verifyAndRestore: function verifyAndRestore() {
@@ -81,15 +54,18 @@
         },
 
         add: function add(fake) {
-            push.call(getFakes(this), fake);
+            if (!this.fakes) {
+                this.fakes = [];
+            }
+            this.fakes.push(fake);
             return fake;
         },
 
-        spy: function spy() {
-            return this.add(sinon.spy.apply(sinon, arguments));
+        spy: function () {
+            return this.add(globalSinon.spy.apply(null, arguments));
         },
 
-        stub: function stub(object, property, value) {
+        stub: function (object, property, value) {
             if (property) {
                 var original = object[property];
 
@@ -108,7 +84,7 @@
                 }
             }
             if (!property && !!object && typeof object == "object") {
-                var stubbedObj = sinon.stub.apply(sinon, arguments);
+                var stubbedObj = globalSinon.stub.apply(null, arguments);
 
                 for (var prop in stubbedObj) {
                     if (typeof stubbedObj[prop] === "function") {
@@ -119,11 +95,11 @@
                 return stubbedObj;
             }
 
-            return this.add(sinon.stub.apply(sinon, arguments));
+            return this.add(globalSinon.stub.apply(null, arguments));
         },
 
-        mock: function mock() {
-            return this.add(sinon.mock.apply(sinon, arguments));
+        mock: function () {
+            return this.add(globalSinon.mock.apply(null, arguments));
         },
 
         inject: function inject(obj) {
@@ -144,10 +120,4 @@
             return obj;
         }
     };
-
-    if (commonJSModule) {
-        module.exports = collection;
-    } else {
-        sinon.collection = collection;
-    }
-}(typeof sinon == "object" && sinon || null));
+};
