+++ lib/sinon/sandbox.js
@@ -1,9 +1,3 @@
-/**
- * @depend ../sinon.js
- * @depend collection.js
- * @depend util/fake_timers.js
- * @depend util/fake_server_with_clock.js
- */
 /*jslint eqeqeq: false, onevar: false, plusplus: false*/
 /*global require, module*/
 /**
@@ -17,59 +11,58 @@
  */
 "use strict";
 
-if (typeof module !== 'undefined' && module.exports) {
-    var sinon = require("../sinon");
-    sinon.extend(sinon, require("./util/fake_timers"));
+var sinon = require("./sinon");
+var fakeServer = require("./util/fake_server");
+//sinon.extend(sinon, require("./util/fake_timers"));
+
+function exposeValue(sandbox, config, key, value) {
+    if (!value) {
+        return;
+    }
+
+    if (config.injectInto) {
+        config.injectInto[key] = value;
+    } else {
+        sandbox.args.push(value);
+    }
 }
 
-(function () {
-    var push = [].push;
+function prepareSandboxFromConfig(proto, config) {
+    var sandbox = sinon.create(proto);
 
-    function exposeValue(sandbox, config, key, value) {
-        if (!value) {
-            return;
+    if (config.useFakeServer) {
+        if (typeof config.useFakeServer == "object") {
+            sandbox.serverPrototype = config.useFakeServer;
         }
 
-        if (config.injectInto) {
-            config.injectInto[key] = value;
+        sandbox.useFakeServer();
+    }
+
+    if (config.useFakeTimers) {
+        if (typeof config.useFakeTimers == "object") {
+            sandbox.useFakeTimers.apply(sandbox, config.useFakeTimers);
         } else {
-            push.call(sandbox.args, value);
+            sandbox.useFakeTimers();
         }
     }
 
-    function prepareSandboxFromConfig(config) {
-        var sandbox = sinon.create(sinon.sandbox);
+    return sandbox;
+}
 
-        if (config.useFakeServer) {
-            if (typeof config.useFakeServer == "object") {
-                sandbox.serverPrototype = config.useFakeServer;
-            }
+module.exports = function (globalSinon) {
 
-            sandbox.useFakeServer();
-        }
+    var proto = sinon.extend(sinon.create(globalSinon.collection), {
 
-        if (config.useFakeTimers) {
-            if (typeof config.useFakeTimers == "object") {
-                sandbox.useFakeTimers.apply(sandbox, config.useFakeTimers);
-            } else {
-                sandbox.useFakeTimers();
-            }
-        }
-
-        return sandbox;
-    }
-
-    sinon.sandbox = sinon.extend(sinon.create(sinon.collection), {
         useFakeTimers: function useFakeTimers() {
-            this.clock = sinon.useFakeTimers.apply(sinon, arguments);
+            this.clock = globalSinon.useFakeTimers.apply(null, arguments);
 
             return this.add(this.clock);
         },
 
-        serverPrototype: sinon.fakeServer,
+        serverPrototype: fakeServer,
 
         useFakeServer: function useFakeServer() {
-            var proto = this.serverPrototype || sinon.fakeServer;
+            var proto = this.serverPrototype || fakeServer;
 
             if (!proto || !proto.create) {
                 return null;
@@ -80,7 +73,7 @@ if (typeof module !== 'undefined' && module.exports) {
         },
 
         inject: function (obj) {
-            sinon.collection.inject.call(this, obj);
+            globalSinon.collection.inject.call(this, obj);
 
             if (this.clock) {
                 obj.clock = this.clock;
@@ -96,10 +89,10 @@ if (typeof module !== 'undefined' && module.exports) {
 
         create: function (config) {
             if (!config) {
-                return sinon.create(sinon.sandbox);
+                return sinon.create(proto);
             }
 
-            var sandbox = prepareSandboxFromConfig(config);
+            var sandbox = prepareSandboxFromConfig(this, config);
             sandbox.args = sandbox.args || [];
             var prop, value, exposed = sandbox.inject({});
 
@@ -115,11 +108,10 @@ if (typeof module !== 'undefined' && module.exports) {
 
             return sandbox;
         }
+
     });
 
-    sinon.sandbox.useFakeXMLHttpRequest = sinon.sandbox.useFakeServer;
+    proto.useFakeXMLHttpRequest = proto.useFakeServer;
 
-    if (typeof module !== 'undefined' && module.exports) {
-        module.exports = sinon.sandbox;
-    }
-}());
+    return proto;
+};
