+++ lib/sinon/stub.js
@@ -1,10 +1,5 @@
-/**
- * @depend ../sinon.js
- * @depend spy.js
- * @depend behavior.js
- */
 /*jslint eqeqeq: false, onevar: false*/
-/*global module, require, sinon*/
+/*global module, require*/
 /**
  * Stub functions
  *
@@ -15,147 +10,129 @@
  */
 "use strict";
 
-(function (sinon) {
-    var commonJSModule = typeof module !== 'undefined' && module.exports;
+var sinon = require("./sinon");
+var spy = require("./spy");
+var behavior = require("./behavior");
 
-    if (!sinon && commonJSModule) {
-        sinon = require("../sinon");
+function stub(object, property, func) {
+    if (!!func && typeof func != "function") {
+        throw new TypeError("Custom stub should be function");
     }
 
-    if (!sinon) {
-        return;
+    var wrapper = func ? spy.create(func) : stub.create();
+
+    if (!object && !property) {
+        return stub.create();
     }
 
-    function stub(object, property, func) {
-        if (!!func && typeof func != "function") {
-            throw new TypeError("Custom stub should be function");
+    if (!property && !!object && typeof object == "object") {
+        for (var prop in object) {
+            if (typeof object[prop] === "function") {
+                stub(object, prop);
+            }
         }
 
-        var wrapper;
+        return object;
+    }
 
-        if (func) {
-            wrapper = sinon.spy && sinon.spy.create ? sinon.spy.create(func) : func;
-        } else {
-            wrapper = stub.create();
-        }
+    return sinon.wrapMethod(object, property, wrapper);
+}
 
-        if (!object && !property) {
-            return sinon.stub.create();
-        }
+function getDefaultBehavior(stub) {
+    return stub.defaultBehavior || getParentBehaviour(stub) || behavior.create(stub);
+}
 
-        if (!property && !!object && typeof object == "object") {
-            for (var prop in object) {
-                if (typeof object[prop] === "function") {
-                    stub(object, prop);
+function getParentBehaviour(stub) {
+    return (stub.parent && getCurrentBehavior(stub.parent));
+}
+
+function getCurrentBehavior(stub) {
+    var behavior = stub.behaviors[stub.callCount - 1];
+    return behavior && behavior.isPresent() ? behavior : getDefaultBehavior(stub);
+}
+
+var uuid = 0;
+
+sinon.extend(stub, (function () {
+    var slice = Array.prototype.slice, proto;
+
+    proto = {
+        create: function create() {
+            var functionStub = function () {
+                return getCurrentBehavior(functionStub).invoke(this, arguments);
+            };
+
+            functionStub.id = "stub#" + uuid++;
+            var orig = functionStub;
+            functionStub = spy.create(functionStub);
+            functionStub.func = orig;
+
+            sinon.extend(functionStub, stub);
+            functionStub._create = stub.create;
+            functionStub.displayName = "stub";
+            functionStub.toString = sinon.functionToString;
+
+            functionStub.defaultBehavior = null;
+            functionStub.behaviors = [];
+
+            return functionStub;
+        },
+
+        resetBehavior: function () {
+            var i;
+
+            this.defaultBehavior = null;
+            this.behaviors = [];
+
+            delete this.returnValue;
+            delete this.returnArgAt;
+            this.returnThis = false;
+
+            if (this.fakes) {
+                for (i = 0; i < this.fakes.length; i++) {
+                    this.fakes[i].resetBehavior();
                 }
             }
+        },
 
-            return object;
+        onCall: function(index) {
+            if (!this.behaviors[index]) {
+                this.behaviors[index] = behavior.create(this);
+            }
+
+            return this.behaviors[index];
+        },
+
+        onFirstCall: function() {
+            return this.onCall(0);
+        },
+
+        onSecondCall: function() {
+            return this.onCall(1);
+        },
+
+        onThirdCall: function() {
+            return this.onCall(2);
         }
+    };
 
-        return sinon.wrapMethod(object, property, wrapper);
-    }
-
-    function getDefaultBehavior(stub) {
-        return stub.defaultBehavior || getParentBehaviour(stub) || sinon.behavior.create(stub);
-    }
-
-    function getParentBehaviour(stub) {
-        return (stub.parent && getCurrentBehavior(stub.parent));
-    }
-
-    function getCurrentBehavior(stub) {
-        var behavior = stub.behaviors[stub.callCount - 1];
-        return behavior && behavior.isPresent() ? behavior : getDefaultBehavior(stub);
-    }
-
-    var uuid = 0;
-
-    sinon.extend(stub, (function () {
-        var slice = Array.prototype.slice, proto;
-
-        proto = {
-            create: function create() {
-                var functionStub = function () {
-                    return getCurrentBehavior(functionStub).invoke(this, arguments);
+    for (var method in behavior) {
+        if (behavior.hasOwnProperty(method) &&
+            !proto.hasOwnProperty(method) &&
+            method != 'create' &&
+            method != 'withArgs' &&
+            method != 'invoke') {
+            proto[method] = (function(behaviorMethod) {
+                return function() {
+                    this.defaultBehavior = this.defaultBehavior || behavior.create(this);
+                    this.defaultBehavior[behaviorMethod].apply(this.defaultBehavior, arguments);
+                    return this;
                 };
-
-                functionStub.id = "stub#" + uuid++;
-                var orig = functionStub;
-                functionStub = sinon.spy.create(functionStub);
-                functionStub.func = orig;
-
-                sinon.extend(functionStub, stub);
-                functionStub._create = sinon.stub.create;
-                functionStub.displayName = "stub";
-                functionStub.toString = sinon.functionToString;
-
-                functionStub.defaultBehavior = null;
-                functionStub.behaviors = [];
-
-                return functionStub;
-            },
-
-            resetBehavior: function () {
-                var i;
-
-                this.defaultBehavior = null;
-                this.behaviors = [];
-
-                delete this.returnValue;
-                delete this.returnArgAt;
-                this.returnThis = false;
-
-                if (this.fakes) {
-                    for (i = 0; i < this.fakes.length; i++) {
-                        this.fakes[i].resetBehavior();
-                    }
-                }
-            },
-
-            onCall: function(index) {
-                if (!this.behaviors[index]) {
-                    this.behaviors[index] = sinon.behavior.create(this);
-                }
-
-                return this.behaviors[index];
-            },
-
-            onFirstCall: function() {
-                return this.onCall(0);
-            },
-
-            onSecondCall: function() {
-                return this.onCall(1);
-            },
-
-            onThirdCall: function() {
-                return this.onCall(2);
-            }
-        };
-
-        for (var method in sinon.behavior) {
-            if (sinon.behavior.hasOwnProperty(method) &&
-                !proto.hasOwnProperty(method) &&
-                method != 'create' &&
-                method != 'withArgs' &&
-                method != 'invoke') {
-                proto[method] = (function(behaviorMethod) {
-                    return function() {
-                        this.defaultBehavior = this.defaultBehavior || sinon.behavior.create(this);
-                        this.defaultBehavior[behaviorMethod].apply(this.defaultBehavior, arguments);
-                        return this;
-                    };
-                }(method));
-            }
+            }(method));
         }
-
-        return proto;
-    }()));
-
-    if (commonJSModule) {
-        module.exports = stub;
-    } else {
-        sinon.stub = stub;
     }
-}(typeof sinon == "object" && sinon || null));
+
+    return proto;
+}()));
+
+module.exports = stub;
