+++ lib/sinon/test.js
@@ -1,11 +1,5 @@
-/**
- * @depend ../sinon.js
- * @depend stub.js
- * @depend mock.js
- * @depend sandbox.js
- */
 /*jslint eqeqeq: false, onevar: false, forin: true, plusplus: false*/
-/*global module, require, sinon*/
+/*global module, require*/
 /**
  * Test function, sandboxes fakes
  *
@@ -16,16 +10,9 @@
  */
 "use strict";
 
-(function (sinon) {
-    var commonJSModule = typeof module !== 'undefined' && module.exports;
+var sinon = require("./sinon");
 
-    if (!sinon && commonJSModule) {
-        sinon = require("../sinon");
-    }
-
-    if (!sinon) {
-        return;
-    }
+module.exports = function (sandbox) {
 
     function test(callback) {
         var type = typeof callback;
@@ -37,9 +24,9 @@
         return function () {
             var config = sinon.getConfig(sinon.config);
             config.injectInto = config.injectIntoThis && this || config.injectInto;
-            var sandbox = sinon.sandbox.create(config);
+            var box = sandbox.create(config);
             var exception, result;
-            var args = Array.prototype.slice.call(arguments).concat(sandbox.args);
+            var args = Array.prototype.slice.call(arguments).concat(box.args);
 
             try {
                 result = callback.apply(this, args);
@@ -48,11 +35,11 @@
             }
 
             if (typeof exception !== "undefined") {
-                sandbox.restore();
+                box.restore();
                 throw exception;
             }
             else {
-                sandbox.verifyAndRestore();
+                box.verifyAndRestore();
             }
 
             return result;
@@ -67,9 +54,5 @@
         useFakeServer: true
     };
 
-    if (commonJSModule) {
-        module.exports = test;
-    } else {
-        sinon.test = test;
-    }
-}(typeof sinon == "object" && sinon || null));
+    return test;
+};
