+++ lib/sinon/test_case.js
@@ -1,9 +1,5 @@
-/**
- * @depend ../sinon.js
- * @depend test.js
- */
 /*jslint eqeqeq: false, onevar: false, eqeqeq: false*/
-/*global module, require, sinon*/
+/*global module, require*/
 /**
  * Test case, sandboxes all test functions
  *
@@ -14,42 +10,33 @@
  */
 "use strict";
 
-(function (sinon) {
-    var commonJSModule = typeof module !== 'undefined' && module.exports;
+function createTest(property, setUp, tearDown) {
+    return function () {
+        if (setUp) {
+            setUp.apply(this, arguments);
+        }
 
-    if (!sinon && commonJSModule) {
-        sinon = require("../sinon");
-    }
+        var exception, result;
 
-    if (!sinon || !Object.prototype.hasOwnProperty) {
-        return;
-    }
+        try {
+            result = property.apply(this, arguments);
+        } catch (e) {
+            exception = e;
+        }
 
-    function createTest(property, setUp, tearDown) {
-        return function () {
-            if (setUp) {
-                setUp.apply(this, arguments);
-            }
+        if (tearDown) {
+            tearDown.apply(this, arguments);
+        }
 
-            var exception, result;
+        if (exception) {
+            throw exception;
+        }
 
-            try {
-                result = property.apply(this, arguments);
-            } catch (e) {
-                exception = e;
-            }
+        return result;
+    };
+}
 
-            if (tearDown) {
-                tearDown.apply(this, arguments);
-            }
-
-            if (exception) {
-                throw exception;
-            }
-
-            return result;
-        };
-    }
+module.exports = function (globalSinon) {
 
     function testCase(tests, prefix) {
         /*jsl:ignore*/
@@ -79,7 +66,7 @@
                         method = createTest(property, setUp, tearDown);
                     }
 
-                    methods[testName] = sinon.test(method);
+                    methods[testName] = globalSinon.test(method);
                 } else {
                     methods[testName] = tests[testName];
                 }
@@ -88,10 +75,6 @@
 
         return methods;
     }
+    return testCase;
 
-    if (commonJSModule) {
-        module.exports = testCase;
-    } else {
-        sinon.testCase = testCase;
-    }
-}(typeof sinon == "object" && sinon || null));
+};
