+++ lib/sinon/util/event.js
@@ -1,5 +1,5 @@
 /*jslint eqeqeq: false, onevar: false*/
-/*global sinon, module, require, ActiveXObject, XMLHttpRequest, DOMParser*/
+/*global module, require*/
 /**
  * Minimal Event interface implementation
  *
@@ -14,62 +14,57 @@
  */
 "use strict";
 
-if (typeof sinon == "undefined") {
-    this.sinon = {};
-}
+function Event(type, bubbles, cancelable, target) {
+    this.initEvent(type, bubbles, cancelable, target);
+};
 
-(function () {
-    var push = [].push;
+Event.prototype = {
+    initEvent: function(type, bubbles, cancelable, target) {
+        this.type = type;
+        this.bubbles = bubbles;
+        this.cancelable = cancelable;
+        this.target = target;
+    },
 
-    sinon.Event = function Event(type, bubbles, cancelable, target) {
-        this.initEvent(type, bubbles, cancelable, target);
-    };
+    stopPropagation: function () {},
 
-    sinon.Event.prototype = {
-        initEvent: function(type, bubbles, cancelable, target) {
-            this.type = type;
-            this.bubbles = bubbles;
-            this.cancelable = cancelable;
-            this.target = target;
-        },
+    preventDefault: function () {
+        this.defaultPrevented = true;
+    }
+};
 
-        stopPropagation: function () {},
+var EventTarget = {
+    addEventListener: function addEventListener(event, listener, useCapture) {
+        this.eventListeners = this.eventListeners || {};
+        this.eventListeners[event] = this.eventListeners[event] || [];
+        this.eventListeners[event].push(listener);
+    },
 
-        preventDefault: function () {
-            this.defaultPrevented = true;
-        }
-    };
+    removeEventListener: function removeEventListener(event, listener, useCapture) {
+        var listeners = this.eventListeners && this.eventListeners[event] || [];
 
-    sinon.EventTarget = {
-        addEventListener: function addEventListener(event, listener, useCapture) {
-            this.eventListeners = this.eventListeners || {};
-            this.eventListeners[event] = this.eventListeners[event] || [];
-            push.call(this.eventListeners[event], listener);
-        },
-
-        removeEventListener: function removeEventListener(event, listener, useCapture) {
-            var listeners = this.eventListeners && this.eventListeners[event] || [];
-
-            for (var i = 0, l = listeners.length; i < l; ++i) {
-                if (listeners[i] == listener) {
-                    return listeners.splice(i, 1);
-                }
+        for (var i = 0, l = listeners.length; i < l; ++i) {
+            if (listeners[i] == listener) {
+                return listeners.splice(i, 1);
             }
-        },
-
-        dispatchEvent: function dispatchEvent(event) {
-            var type = event.type;
-            var listeners = this.eventListeners && this.eventListeners[type] || [];
-
-            for (var i = 0; i < listeners.length; i++) {
-                if (typeof listeners[i] == "function") {
-                    listeners[i].call(this, event);
-                } else {
-                    listeners[i].handleEvent(event);
-                }
-            }
-
-            return !!event.defaultPrevented;
         }
-    };
-}());
+    },
+
+    dispatchEvent: function dispatchEvent(event) {
+        var type = event.type;
+        var listeners = this.eventListeners && this.eventListeners[type] || [];
+
+        for (var i = 0; i < listeners.length; i++) {
+            if (typeof listeners[i] == "function") {
+                listeners[i].call(this, event);
+            } else {
+                listeners[i].handleEvent(event);
+            }
+        }
+
+        return !!event.defaultPrevented;
+    }
+};
+
+exports.Event = Event;
+exports.EventTarget = EventTarget;
