+++ lib/sinon/util/fake_server.js
@@ -1,6 +1,3 @@
-/**
- * @depend fake_xml_http_request.js
- */
 /*jslint eqeqeq: false, onevar: false, regexp: false, plusplus: false*/
 /*global module, require, window*/
 /**
@@ -16,197 +13,180 @@
  */
 "use strict";
 
-if (typeof sinon == "undefined") {
-    var sinon = {};
+var sinon = require("../sinon");
+var push = Array.prototype.push;
+
+function responseArray(handler) {
+    var response = handler;
+
+    if (Object.prototype.toString.call(handler) != "[object Array]") {
+        response = [200, {}, handler];
+    }
+
+    if (typeof response[2] != "string") {
+        throw new TypeError("Fake server response body should be string, but was " +
+                            typeof response[2]);
+    }
+
+    return response;
 }
 
-sinon.fakeServer = (function () {
-    var push = [].push;
-    function F() {}
+var wloc = typeof window !== "undefined" ? window.location : {};
+var rCurrLoc = new RegExp("^" + wloc.protocol + "//" + wloc.host);
 
-    function create(proto) {
-        F.prototype = proto;
-        return new F();
+function matchOne(response, reqMethod, reqUrl) {
+    var rmeth = response.method;
+    var matchMethod = !rmeth || rmeth.toLowerCase() == reqMethod.toLowerCase();
+    var url = response.url;
+    var matchUrl = !url || url == reqUrl || (typeof url.test == "function" && url.test(reqUrl));
+
+    return matchMethod && matchUrl;
+}
+
+function match(response, request) {
+    var requestMethod = this.getHTTPMethod(request);
+    var requestUrl = request.url;
+
+    if (!/^https?:\/\//.test(requestUrl) || rCurrLoc.test(requestUrl)) {
+        requestUrl = requestUrl.replace(rCurrLoc, "");
     }
 
-    function responseArray(handler) {
-        var response = handler;
-
-        if (Object.prototype.toString.call(handler) != "[object Array]") {
-            response = [200, {}, handler];
+    if (matchOne(response, this.getHTTPMethod(request), requestUrl)) {
+        if (typeof response.response == "function") {
+            var ru = response.url;
+            var args = [request].concat(ru && typeof ru.exec == "function" ? ru.exec(requestUrl).slice(1) : []);
+            return response.response.apply(response, args);
         }
 
-        if (typeof response[2] != "string") {
-            throw new TypeError("Fake server response body should be string, but was " +
-                                typeof response[2]);
-        }
-
-        return response;
+        return true;
     }
 
-    var wloc = typeof window !== "undefined" ? window.location : {};
-    var rCurrLoc = new RegExp("^" + wloc.protocol + "//" + wloc.host);
+    return false;
+}
 
-    function matchOne(response, reqMethod, reqUrl) {
-        var rmeth = response.method;
-        var matchMethod = !rmeth || rmeth.toLowerCase() == reqMethod.toLowerCase();
-        var url = response.url;
-        var matchUrl = !url || url == reqUrl || (typeof url.test == "function" && url.test(reqUrl));
+function log(response, request) {
+    var str;
 
-        return matchMethod && matchUrl;
-    }
+    str =  "Request:\n"  + sinon.format(request)  + "\n\n";
+    str += "Response:\n" + sinon.format(response) + "\n\n";
 
-    function match(response, request) {
-        var requestMethod = this.getHTTPMethod(request);
-        var requestUrl = request.url;
+    sinon.log(str);
+}
 
-        if (!/^https?:\/\//.test(requestUrl) || rCurrLoc.test(requestUrl)) {
-            requestUrl = requestUrl.replace(rCurrLoc, "");
-        }
+exports.create = function () {
+    var server = sinon.create(this);
+    this.xhr = sinon.useFakeXMLHttpRequest();
+    server.requests = [];
 
-        if (matchOne(response, this.getHTTPMethod(request), requestUrl)) {
-            if (typeof response.response == "function") {
-                var ru = response.url;
-                var args = [request].concat(ru && typeof ru.exec == "function" ? ru.exec(requestUrl).slice(1) : []);
-                return response.response.apply(response, args);
-            }
+    this.xhr.onCreate = function (xhrObj) {
+        server.addRequest(xhrObj);
+    };
 
-            return true;
-        }
+    return server;
+};
 
-        return false;
-    }
+exports.addRequest = function addRequest(xhrObj) {
+    var server = this;
+    push.call(this.requests, xhrObj);
 
-    function log(response, request) {
-        var str;
+    xhrObj.onSend = function () {
+        server.handleRequest(this);
 
-        str =  "Request:\n"  + sinon.format(request)  + "\n\n";
-        str += "Response:\n" + sinon.format(response) + "\n\n";
+        if (server.autoRespond && !server.responding) {
+            setTimeout(function () {
+                server.responding = false;
+                server.respond();
+            }, server.autoRespondAfter || 10);
 
-        sinon.log(str);
-    }
-
-    return {
-        create: function () {
-            var server = create(this);
-            this.xhr = sinon.useFakeXMLHttpRequest();
-            server.requests = [];
-
-            this.xhr.onCreate = function (xhrObj) {
-                server.addRequest(xhrObj);
-            };
-
-            return server;
-        },
-
-        addRequest: function addRequest(xhrObj) {
-            var server = this;
-            push.call(this.requests, xhrObj);
-
-            xhrObj.onSend = function () {
-                server.handleRequest(this);
-
-                if (server.autoRespond && !server.responding) {
-                    setTimeout(function () {
-                        server.responding = false;
-                        server.respond();
-                    }, server.autoRespondAfter || 10);
-
-                    server.responding = true;
-                }
-            };
-        },
-
-        getHTTPMethod: function getHTTPMethod(request) {
-            if (this.fakeHTTPMethods && /post/i.test(request.method)) {
-                var matches = (request.requestBody || "").match(/_method=([^\b;]+)/);
-                return !!matches ? matches[1] : request.method;
-            }
-
-            return request.method;
-        },
-
-        handleRequest: function handleRequest(xhr) {
-            if (xhr.async) {
-                if (!this.queue) {
-                    this.queue = [];
-                }
-
-                push.call(this.queue, xhr);
-            } else {
-                this.processRequest(xhr);
-            }
-        },
-
-        respondWith: function respondWith(method, url, body) {
-            if (arguments.length == 1 && typeof method != "function") {
-                this.response = responseArray(method);
-                return;
-            }
-
-            if (!this.responses) { this.responses = []; }
-
-            if (arguments.length == 1) {
-                body = method;
-                url = method = null;
-            }
-
-            if (arguments.length == 2) {
-                body = url;
-                url = method;
-                method = null;
-            }
-
-            push.call(this.responses, {
-                method: method,
-                url: url,
-                response: typeof body == "function" ? body : responseArray(body)
-            });
-        },
-
-        respond: function respond() {
-            if (arguments.length > 0) this.respondWith.apply(this, arguments);
-            var queue = this.queue || [];
-            var request;
-
-            while(request = queue.shift()) {
-                this.processRequest(request);
-            }
-        },
-
-        processRequest: function processRequest(request) {
-            try {
-                if (request.aborted) {
-                    return;
-                }
-
-                var response = this.response || [404, {}, ""];
-
-                if (this.responses) {
-                    for (var i = 0, l = this.responses.length; i < l; i++) {
-                        if (match.call(this, this.responses[i], request)) {
-                            response = this.responses[i].response;
-                            break;
-                        }
-                    }
-                }
-
-                if (request.readyState != 4) {
-                    log(response, request);
-
-                    request.respond(response[0], response[1], response[2]);
-                }
-            } catch (e) {
-                sinon.logError("Fake server request processing", e);
-            }
-        },
-
-        restore: function restore() {
-            return this.xhr.restore && this.xhr.restore.apply(this.xhr, arguments);
+            server.responding = true;
         }
     };
-}());
+};
 
-if (typeof module !== 'undefined' && module.exports) {
-    module.exports = sinon;
-}
+exports.getHTTPMethod = function getHTTPMethod(request) {
+    if (this.fakeHTTPMethods && /post/i.test(request.method)) {
+        var matches = (request.requestBody || "").match(/_method=([^\b;]+)/);
+        return !!matches ? matches[1] : request.method;
+    }
+
+    return request.method;
+};
+
+exports.handleRequest = function handleRequest(xhr) {
+    if (xhr.async) {
+        if (!this.queue) {
+            this.queue = [];
+        }
+
+        push.call(this.queue, xhr);
+    } else {
+        this.processRequest(xhr);
+    }
+};
+
+exports.respondWith = function respondWith(method, url, body) {
+    if (arguments.length == 1 && typeof method != "function") {
+        this.response = responseArray(method);
+        return;
+    }
+
+    if (!this.responses) { this.responses = []; }
+
+    if (arguments.length == 1) {
+        body = method;
+        url = method = null;
+    }
+
+    if (arguments.length == 2) {
+        body = url;
+        url = method;
+        method = null;
+    }
+
+    push.call(this.responses, {
+        method: method,
+        url: url,
+        response: typeof body == "function" ? body : responseArray(body)
+    });
+};
+
+exports.respond = function respond() {
+    if (arguments.length > 0) this.respondWith.apply(this, arguments);
+    var queue = this.queue || [];
+    var request;
+
+    while(request = queue.shift()) {
+        this.processRequest(request);
+    }
+};
+
+exports.processRequest = function processRequest(request) {
+    try {
+        if (request.aborted) {
+            return;
+        }
+
+        var response = this.response || [404, {}, ""];
+
+        if (this.responses) {
+            for (var i = 0, l = this.responses.length; i < l; i++) {
+                if (match.call(this, this.responses[i], request)) {
+                    response = this.responses[i].response;
+                    break;
+                }
+            }
+        }
+
+        if (request.readyState != 4) {
+            log(response, request);
+
+            request.respond(response[0], response[1], response[2]);
+        }
+    } catch (e) {
+        sinon.logError("Fake server request processing", e);
+    }
+};
+
+exports.restore = function restore() {
+    return this.xhr.restore && this.xhr.restore.apply(this.xhr, arguments);
+};
