+++ lib/sinon/util/fake_server_with_clock.js
@@ -3,7 +3,7 @@
  * @depend fake_timers.js
  */
 /*jslint browser: true, eqeqeq: false, onevar: false*/
-/*global sinon*/
+/*global*/
 /**
  * Add-on for sinon.fakeServer that automatically handles a fake timer along with
  * the FakeXMLHttpRequest. The direct inspiration for this add-on is jQuery
@@ -20,64 +20,67 @@
  */
 "use strict";
 
-(function () {
-    function Server() {}
-    Server.prototype = sinon.fakeServer;
+var fakeServer = require("./fake_server");
+var fakeTimers = require("./fake_timers");
 
-    sinon.fakeServerWithClock = new Server();
+function Server() {}
+Server.prototype = fakeServer;
 
-    sinon.fakeServerWithClock.addRequest = function addRequest(xhr) {
-        if (xhr.async) {
-            if (typeof setTimeout.clock == "object") {
-                this.clock = setTimeout.clock;
-            } else {
-                this.clock = sinon.useFakeTimers();
-                this.resetClock = true;
-            }
+var fakeServerWithClock = new Server();
 
-            if (!this.longestTimeout) {
-                var clockSetTimeout = this.clock.setTimeout;
-                var clockSetInterval = this.clock.setInterval;
-                var server = this;
-
-                this.clock.setTimeout = function (fn, timeout) {
-                    server.longestTimeout = Math.max(timeout, server.longestTimeout || 0);
-
-                    return clockSetTimeout.apply(this, arguments);
-                };
-
-                this.clock.setInterval = function (fn, timeout) {
-                    server.longestTimeout = Math.max(timeout, server.longestTimeout || 0);
-
-                    return clockSetInterval.apply(this, arguments);
-                };
-            }
+fakeServerWithClock.addRequest = function addRequest(xhr) {
+    if (xhr.async) {
+        if (typeof setTimeout.clock == "object") {
+            this.clock = setTimeout.clock;
+        } else {
+            this.clock = fakeTimers.useFakeTimers();
+            this.resetClock = true;
         }
 
-        return sinon.fakeServer.addRequest.call(this, xhr);
-    };
+        if (!this.longestTimeout) {
+            var clockSetTimeout = this.clock.setTimeout;
+            var clockSetInterval = this.clock.setInterval;
+            var server = this;
 
-    sinon.fakeServerWithClock.respond = function respond() {
-        var returnVal = sinon.fakeServer.respond.apply(this, arguments);
+            this.clock.setTimeout = function (fn, timeout) {
+                server.longestTimeout = Math.max(timeout, server.longestTimeout || 0);
 
-        if (this.clock) {
-            this.clock.tick(this.longestTimeout || 0);
-            this.longestTimeout = 0;
+                return clockSetTimeout.apply(this, arguments);
+            };
 
-            if (this.resetClock) {
-                this.clock.restore();
-                this.resetClock = false;
-            }
+            this.clock.setInterval = function (fn, timeout) {
+                server.longestTimeout = Math.max(timeout, server.longestTimeout || 0);
+
+                return clockSetInterval.apply(this, arguments);
+            };
         }
+    }
 
-        return returnVal;
-    };
+    return fakeServer.addRequest.call(this, xhr);
+};
 
-    sinon.fakeServerWithClock.restore = function restore() {
-        if (this.clock) {
+fakeServerWithClock.respond = function respond() {
+    var returnVal = fakeServer.respond.apply(this, arguments);
+
+    if (this.clock) {
+        this.clock.tick(this.longestTimeout || 0);
+        this.longestTimeout = 0;
+
+        if (this.resetClock) {
             this.clock.restore();
+            this.resetClock = false;
         }
+    }
 
-        return sinon.fakeServer.restore.apply(this, arguments);
-    };
-}());
+    return returnVal;
+};
+
+fakeServerWithClock.restore = function restore() {
+    if (this.clock) {
+        this.clock.restore();
+    }
+
+    return fakeServer.restore.apply(this, arguments);
+};
+
+module.exports = fakeServerWithClock;
