+++ lib/sinon/util/fake_timers.js
@@ -19,10 +19,6 @@
  */
 "use strict";
 
-if (typeof sinon == "undefined") {
-    var sinon = {};
-}
-
 (function (global) {
     var id = 1;
 
@@ -93,7 +89,7 @@ if (typeof sinon == "undefined") {
         return newObject;
     }
 
-    sinon.clock = {
+    exports.clock = {
         now: 0,
 
         create: function create(now) {
@@ -199,6 +195,8 @@ if (typeof sinon == "undefined") {
                 if (typeof timer.func == "function") {
                     timer.func.apply(null, timer.invokeArgs);
                 } else {
+                    // FIXME This feature is impossible to implement correctly
+                    // because the scope is wrong.
                     eval(timer.func);
                 }
             } catch (e) {
@@ -320,8 +318,8 @@ if (typeof sinon == "undefined") {
         global[method].clock = clock;
     }
 
-    sinon.useFakeTimers = function useFakeTimers(now) {
-        var clock = sinon.clock.create(now);
+    exports.useFakeTimers = function useFakeTimers(now) {
+        var clock = exports.clock.create(now);
         clock.restore = restore;
         clock.methods = Array.prototype.slice.call(arguments,
                                                    typeof now == "number" ? 1 : 0);
@@ -338,14 +336,10 @@ if (typeof sinon == "undefined") {
     };
 }(typeof global != "undefined" && typeof global !== "function" ? global : this));
 
-sinon.timers = {
+exports.timers = {
     setTimeout: setTimeout,
     clearTimeout: clearTimeout,
     setInterval: setInterval,
     clearInterval: clearInterval,
     Date: Date
 };
-
-if (typeof module !== 'undefined' && module.exports) {
-    module.exports = sinon;
-}
