+++ lib/sinon/util/fake_xml_http_request.js
@@ -1,9 +1,5 @@
-/**
- * @depend ../../sinon.js
- * @depend event.js
- */
 /*jslint eqeqeq: false, onevar: false*/
-/*global sinon, module, require, ActiveXObject, XMLHttpRequest, DOMParser*/
+/*global module, require, ActiveXObject, XMLHttpRequest, DOMParser*/
 /**
  * Fake XMLHttpRequest object
  *
@@ -14,15 +10,16 @@
  */
 "use strict";
 
+var sinon = require("../sinon");
+var Event = require("./event").Event;
+var EventTarget = require("./event").EventTarget;
+
 // wrapper for global
 (function(global) {
 
-    if (typeof sinon === "undefined") {
-        global.sinon = {};
-    }
-    sinon.xhr = { XMLHttpRequest: global.XMLHttpRequest };
+    var xhr = { XMLHttpRequest: global.XMLHttpRequest };
+    exports.xhr = xhr;
 
-    var xhr = sinon.xhr;
     xhr.GlobalXMLHttpRequest = global.XMLHttpRequest;
     xhr.GlobalActiveXObject = global.ActiveXObject;
     xhr.supportsActiveX = typeof xhr.GlobalActiveXObject != "undefined";
@@ -150,7 +147,7 @@
     };
     var IE6Re = /MSIE 6/;
     FakeXMLHttpRequest.defake = function(fakeXhr,xhrArgs) {
-        var xhr = new sinon.xhr.workingXHR();
+        var xhr = new xhr.workingXHR();
         each(["open","setRequestHeader","send","abort","getResponseHeader",
               "getAllResponseHeaders","addEventListener","overrideMimeType","removeEventListener"],
              function(method) {
@@ -219,7 +216,7 @@
         }
     }
 
-    sinon.extend(FakeXMLHttpRequest.prototype, sinon.EventTarget, {
+    sinon.extend(FakeXMLHttpRequest.prototype, EventTarget, {
         async: true,
 
         open: function open(method, url, async, username, password) {
@@ -232,13 +229,13 @@
             this.responseXML = null;
             this.requestHeaders = {};
             this.sendFlag = false;
-            if(sinon.FakeXMLHttpRequest.useFilters === true) {
+            if(FakeXMLHttpRequest.useFilters === true) {
                 var xhrArgs = arguments;
                 var defake = some(FakeXMLHttpRequest.filters,function(filter) {
                     return filter.apply(this,xhrArgs)
                 });
                 if (defake) {
-                  return sinon.FakeXMLHttpRequest.defake(this,arguments);
+                  return FakeXMLHttpRequest.defake(this,arguments);
                 }
             }
             this.readyStateChange(FakeXMLHttpRequest.OPENED);
@@ -255,13 +252,13 @@
                 }
             }
 
-            this.dispatchEvent(new sinon.Event("readystatechange"));
+            this.dispatchEvent(new Event("readystatechange"));
 
             switch (this.readyState) {
                 case FakeXMLHttpRequest.DONE:
-                    this.dispatchEvent(new sinon.Event("load", false, false, this));
-                    this.dispatchEvent(new sinon.Event("loadend", false, false, this));
-                    this.upload.dispatchEvent(new sinon.Event("load", false, false, this));
+                    this.dispatchEvent(new Event("load", false, false, this));
+                    this.dispatchEvent(new Event("loadend", false, false, this));
+                    this.upload.dispatchEvent(new Event("load", false, false, this));
                     this.upload.dispatchEvent(new ProgressEvent("progress", {loaded: 100, total: 100}));
                     break;
             }
@@ -321,7 +318,7 @@
                 this.onSend(this);
             }
 
-            this.dispatchEvent(new sinon.Event("loadstart", false, false, this));
+            this.dispatchEvent(new Event("loadstart", false, false, this));
         },
 
         abort: function abort() {
@@ -330,16 +327,16 @@
             this.errorFlag = true;
             this.requestHeaders = {};
 
-            if (this.readyState > sinon.FakeXMLHttpRequest.UNSENT && this.sendFlag) {
-                this.readyStateChange(sinon.FakeXMLHttpRequest.DONE);
+            if (this.readyState > FakeXMLHttpRequest.UNSENT && this.sendFlag) {
+                this.readyStateChange(FakeXMLHttpRequest.DONE);
                 this.sendFlag = false;
             }
 
-            this.readyState = sinon.FakeXMLHttpRequest.UNSENT;
+            this.readyState = FakeXMLHttpRequest.UNSENT;
 
-            this.dispatchEvent(new sinon.Event("abort", false, false, this));
+            this.dispatchEvent(new Event("abort", false, false, this));
 
-            this.upload.dispatchEvent(new sinon.Event("abort", false, false, this));
+            this.upload.dispatchEvent(new Event("abort", false, false, this));
 
             if (typeof this.onerror === "function") {
                 this.onerror();
@@ -503,8 +500,8 @@
         505: "HTTP Version Not Supported"
     };
 
-    sinon.useFakeXMLHttpRequest = function () {
-        sinon.FakeXMLHttpRequest.restore = function restore(keepOnCreate) {
+    exports.useFakeXMLHttpRequest = function () {
+        FakeXMLHttpRequest.restore = function restore(keepOnCreate) {
             if (xhr.supportsXHR) {
                 global.XMLHttpRequest = xhr.GlobalXMLHttpRequest;
             }
@@ -513,34 +510,30 @@
                 global.ActiveXObject = xhr.GlobalActiveXObject;
             }
 
-            delete sinon.FakeXMLHttpRequest.restore;
+            delete FakeXMLHttpRequest.restore;
 
             if (keepOnCreate !== true) {
-                delete sinon.FakeXMLHttpRequest.onCreate;
+                delete FakeXMLHttpRequest.onCreate;
             }
         };
         if (xhr.supportsXHR) {
-            global.XMLHttpRequest = sinon.FakeXMLHttpRequest;
+            global.XMLHttpRequest = FakeXMLHttpRequest;
         }
 
         if (xhr.supportsActiveX) {
             global.ActiveXObject = function ActiveXObject(objId) {
                 if (objId == "Microsoft.XMLHTTP" || /^Msxml2\.XMLHTTP/i.test(objId)) {
 
-                    return new sinon.FakeXMLHttpRequest();
+                    return new FakeXMLHttpRequest();
                 }
 
                 return new xhr.GlobalActiveXObject(objId);
             };
         }
 
-        return sinon.FakeXMLHttpRequest;
+        return FakeXMLHttpRequest;
     };
 
-    sinon.FakeXMLHttpRequest = FakeXMLHttpRequest;
+    exports.FakeXMLHttpRequest = FakeXMLHttpRequest;
 
 })(typeof global === "object" ? global : this);
-
-if (typeof module !== 'undefined' && module.exports) {
-    module.exports = sinon;
-}
