+++ test/sinon.html
@@ -28,6 +28,8 @@
     <!-- Custom runner -->
     <script src="runner.js"></script>
     <!-- Sources -->
+    <script src="sinon-bundle.js"></script>
+    <!--
     <script src="../lib/sinon.js"></script>
     <script src="../lib/sinon/match.js"></script>
     <script src="../lib/sinon/spy.js"></script>
@@ -47,6 +49,7 @@
     <script src="../lib/sinon/sandbox.js"></script>
     <script src="../lib/sinon/test.js"></script>
     <script src="../lib/sinon/test_case.js"></script>
+    -->
     <!-- Tests -->
     <script src="sinon_test.js"></script>
     <script src="sinon/spy_test.js"></script>
